Removed Pedometer require
